+++ NEWS.md
@@ -1,3 +1,13 @@
+Eunomia 2.0
+=============
+Changes
+- Updated package to no longer contain a dataset rather facilitate access to sample datasets
+  stored in the https://github.com/OHDSI/EunomiaDatasets repository
+- Backward compatibility maintained with getEunomiaConnectionDetails function
+- New function added for getDatabaseFile
+- Embedded sample dataset removed
+- Remove dependency on DatabaseConnector and Java
+
 Eunomia 1.0.3
 =============